# plugin/.well-known/openapi.yaml
openapi: 3.0.1
info:
    title: bilibili弹幕分析
    description: 本项目是一个基于bilibili弹幕和评论的分析项目，通过对弹幕的分析，可以得到弹幕中的关键词，以及弹幕的分析结果。
    version: "v1"
servers:
    - url: http://127.0.0.1:8081
paths:
    /analyseDMK:
        post:
            operationId: analyseDMK
            summary: 传入bilibili的BV号或BV链接,并对其中的弹幕（不是评论）进行多方位分析
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/analyseDMK_input"
            responses:
                "200":
                    description: 成功接收到bv号,总结分析弹幕前五十的主要情感趋势以及概括
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/wordResponse"
    /analyseComment:
        post:
            operationId: analyseComment
            summary: 传入bilibili的BV号或BV链接,并对其中的评论（不是弹幕）进行多方位分析
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/analyseComment_input"
            responses:
                "200":
                    description: 返回对其中的评论（不是弹幕）进行的多方位分析
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/commentResponse"
    /analyseBoth:
        post:
            operationId: analyseBoth
            summary: 传入bilibili的BV号或BV链接,并对其中的评论和弹幕进行多方位分析
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/analyseBoth_input"
            responses:
                "200":
                    description: 返回对其中的评论和弹幕进行的多方位分析
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/bothResponse"

components:
    schemas:
        messageResponse:
            type: object
            required: [message]
            properties:
                message:
                    type: string
                    description: 回复信息
        analyseDMK_input:
            type: object
            required: [url]
            properties:
                url:
                    type: String
                    description: BV号或者bilibili_url链接
        analyseComment_input:
            type: object
            required: [url]
            properties:
                url:
                    type: String
                    description: BV号或者bilibili_url链接
        analyseBoth_input:
            type: object
            required: [url]
            properties:
                url:
                    type: String
                    description: BV号或者bilibili_url链接
        wordResponse:
            type: object
            required: [str_list]
            properties:
                str_list:
                    type: array
                    items:
                        type: string
                        description: 弹幕内容
        commentResponse:
            type: object
            required: [str_list]
            properties:
                str_list:
                    type: array
                    items:
                        type: string
                        description: 评论内容
        bothResponse:
            type: object
            required: [str_list]
            properties:
                both_list:
                    type: array
                    items:
                        type: string
                        description: 评论内容